Fix package.
This isn't library package, just executable, so we don't need to copy
libraries, headers, etc. The package_info method is not necessary
because this isn't library and we don't need to publish any type of
dependency information.

Expand describe usage

Tag name can't have '/' as this is delimiter for conan versions. So
replace '/' with '-'.

Add dirty qualifier to version name so we know if it has local
modifications.